OSIS-1150 ExternalAcronym and Organization for external_learning_unit
