+++ base/forms/learning_unit/learning_unit_create_2.py
@@ -29,28 +29,17 @@ from collections import OrderedDict
 from django.db import transaction
 from django.utils.functional import cached_property
 from django.utils.translation import ugettext_lazy as _
-from django import forms
 
 from base.business.utils.model import merge_two_dicts
 from base.forms.learning_unit.learning_unit_create import LearningUnitModelForm, LearningUnitYearModelForm, \
-    LearningContainerModelForm, LearningContainerYearModelForm, EntityContainerBaseForm, LearningContainerYearExternalModelForm
+    LearningContainerModelForm, LearningContainerYearModelForm, EntityContainerBaseForm
 from base.forms.utils.acronym_field import split_acronym
+# from base.models.entity_version import EntityVersion
 from base.models import learning_unit_year
 from base.models.campus import Campus
 from base.models.enums import learning_unit_year_subtypes
 from base.models.learning_unit import LearningUnit
 from reference.models import language
-from base.models.enums import learning_container_year_types
-from base.models.external_learning_unit_year import ExternalLearningUnitYear
-from base.models.entity_version import find_main_entities_version
-from base.models.learning_container_year import LearningContainerYear
-from base.models.enums import learning_container_year_types
-from base.models.enums.learning_container_year_types import EXTERNAL
-from django.forms import ModelChoiceField
-from base.models.campus import find_main_campuses
-from base.models.entity import Entity
-# from base.models.entity_version import EntityVersion
-from base.models import entity_version
 
 FULL_READ_ONLY_FIELDS = {"acronym", "academic_year", "container_type"}
 FULL_PROPOSAL_READ_ONLY_FIELDS = {"academic_year", "container_type"}
@@ -163,6 +152,10 @@ class LearningUnitBaseForm(metaclass=ABCMeta):
                 "common_title", _("must_set_common_title_or_specific_title"))
             return False
         return True
+
+    @property
+    def learning_container_form(self):
+        return self.forms[LearningContainerModelForm]
 
     @property
     def learning_unit_form(self):
@@ -288,35 +281,29 @@ class FullForm(LearningUnitBaseForm):
     def save(self, commit=True):
         academic_year = self.academic_year
 
-        learning_unit_year = save_learning_unit_year(academic_year, commit, self.start_year, self.forms)
+        learning_container = self.forms[LearningContainerModelForm].save(commit)
+        learning_unit = self.forms[LearningUnitModelForm].save(
+            start_year=self.start_year,
+            learning_container=learning_container,
+            commit=commit
+        )
+        container_year = self.forms[LearningContainerYearModelForm].save(
+            academic_year=academic_year,
+            learning_container=learning_container,
+            acronym=self.forms[LearningUnitYearModelForm].instance.acronym,
+            commit=commit
+        )
+
+        entity_container_years = self.entity_container_form.save(commit=commit, learning_container_year=container_year)
+
+        # Save learning unit year (learning_unit_component +  learning_component_year + entity_component_year)
+        learning_unit_year = self.forms[LearningUnitYearModelForm].save(
+            learning_container_year=container_year,
+            learning_unit=learning_unit,
+            entity_container_years=entity_container_years,
+            commit=commit
+        )
         return learning_unit_year
-
-
-def save_learning_unit_year(forms, academic_year, commit, start_year):
-    learning_container = forms[LearningContainerModelForm].save(commit)
-    learning_unit = forms[LearningUnitModelForm].save(
-        start_year=start_year,
-        learning_container=learning_container,
-        commit=commit
-    )
-    container_year = forms[LearningContainerYearModelForm].save(
-        academic_year=academic_year,
-        learning_container=learning_container,
-        acronym=forms[LearningUnitYearModelForm].instance.acronym,
-        commit=commit
-    )
-    if entity_container_form:
-        entity_container_years = entity_container_form.save(commit=commit, learning_container_year=container_year)
-    else:
-        entity_container_years = None
-    # Save learning unit year (learning_unit_component +  learning_component_year + entity_component_year)
-    learning_unit_year = forms[LearningUnitYearModelForm].save(
-        learning_container_year=container_year,
-        learning_unit=learning_unit,
-        entity_container_years=entity_container_years,
-        commit=commit
-    )
-    return learning_unit_year
 
 
 def merge_data(data, inherit_lu_values):
@@ -444,172 +431,3 @@ class PartimForm(LearningUnitBaseForm):
 
     def _get_entity_container_year(self):
         return self.learning_unit_year_full.learning_container_year.entitycontaineryear_set.all()
-
-
-class EntityChoiceField(ModelChoiceField):
-    def label_from_instance(self, obj):
-        return "{0}".format(obj.acronym)
-
-    def __init__(self, *args, **kwargs):
-        super(EntityChoiceField, self).__init__(*args, **kwargs)
-        self.widget.attrs['class'] = 'form-control'
-
-
-class CampusChoiceField(ModelChoiceField):
-    def label_from_instance(self, obj):
-        return "{0} ({1})".format(obj.name, obj.organization.name)
-
-    def __init__(self, *args, **kwargs):
-        super(CampusChoiceField, self).__init__(*args, **kwargs)
-        self.widget.attrs['class'] = 'form-control'
-
-
-class LearningUnitExternalModelForm(forms.ModelForm):
-    buyer = EntityChoiceField(required=True, queryset=None)
-    entity_version = None
-
-    def __init__(self, *args, **kwargs):
-        self.person = kwargs.pop('person')
-        super().__init__(*args, **kwargs)
-        field = self.fields['buyer']
-        field.queryset = self.person.find_main_entities_version
-        field.label =_('requesting_entity')
-        self.fields['external_credits'].label = _('local_credits')
-
-        if hasattr(self.instance, 'buyer'):
-            self.initial['buyer'] = get_last_version(self.instance.buyer)
-
-    class Meta:
-        model = ExternalLearningUnitYear
-        fields = ('external_acronym', 'external_credits', 'url', 'learning_unit_year', 'buyer')
-
-    def clean_buyer(self):
-        ev_data = self.cleaned_data['buyer']
-        self.entity_version = ev_data
-        return ev_data.entity if ev_data else None
-
-    def post_clean(self, start_date):
-        entity = self.cleaned_data.get('buyer')
-        if not entity:
-            return
-        entity_v= entity_version.get_by_entity_and_date(entity, start_date)
-        if not entity_v:
-            self.add_error('buyer', _("The linked entity does not exist at the start date of the "
-                                       "academic year linked to this learning unit"))
-        else:
-            self.entity_version=entity_v
-
-
-class LearningUnitExternalForm(LearningUnitBaseForm):
-    forms = OrderedDict()
-    academic_year = None
-    subtype = learning_unit_year_subtypes.FULL
-    container_type = ((EXTERNAL, _(EXTERNAL)),)
-    entity_version = None
-
-    def __init__(self, person, academic_year=None, data=None, *args, **kwargs):
-        self.form_cls_to_validate = [
-            LearningUnitModelForm,
-            LearningUnitYearModelForm,
-            LearningContainerModelForm,
-            LearningContainerYearExternalModelForm,
-            LearningUnitExternalModelForm
-        ]
-        self.academic_year = academic_year
-        self.person = person
-        instances_data = self._build_instance_data(data, academic_year, None)
-        super().__init__(instances_data, *args, **kwargs)
-        self.fields['campus'].queryset = find_main_campuses().order_by('organization__name')
-        self.fields['container_type'] = (('EXTERNAL','EXTERNAL'),)
-
-    @property
-    def learning_unit_external_form(self):
-        return self.forms.get(LearningUnitExternalModelForm, None)
-
-    @property
-    def learning_container_year_external_form(self):
-        return self.forms.get(LearningContainerYearExternalModelForm, None)
-
-    def _build_instance_data(self, data, default_ac_year, proposal):
-        return {
-            LearningUnitModelForm: {
-                'data': merge_data(data, {'start_year': default_ac_year.year, 'periodicity': 'ANNUAL'}),
-                'instance': self.instance.learning_unit if self.instance else None,
-            },
-            LearningContainerModelForm: {
-                'data': data,
-                'instance': self.instance.learning_container_year.learning_container if self.instance else None,
-            },
-            LearningUnitYearModelForm: self._build_instance_data_learning_unit_year(data, default_ac_year),
-            LearningContainerYearExternalModelForm: self._build_instance_data_learning_container_year(data, proposal),
-            LearningUnitExternalModelForm: self._build_instance_data_external_learning_unit(default_ac_year,data)
-        }
-
-    def _build_instance_data_external_learning_unit(self, academic_year, data):
-        return {
-            'data': data,
-            'instance': None,
-            'initial': {'external_credits': None, 'url': None, 'buyer': None, 'external_acronym': ''},
-            'person': self.person
-        }
-
-    def _build_instance_data_learning_unit_year(self, data, default_ac_year):
-
-        return {
-            'data': merge_data(data, {'periodicity': 'ANNUAL'}),
-            'instance': self.instance,
-            'initial': {'status':True,
-                        'academic_year': default_ac_year},
-            'person': self.person,
-            'subtype': self.subtype
-        }
-
-    def _build_instance_data_learning_unit(self, data, inherit_lu_values):
-        return {
-            'data': merge_data(data, inherit_lu_values),
-            'instance': None,
-            'initial': inherit_lu_values if not self.instance else None,
-        }
-
-    def _build_instance_data_learning_container_year(self, data, proposal):
-        container_type=((EXTERNAL, _(EXTERNAL)),)
-        return {
-            'data': data,
-            'instance': self.instance.learning_container_year if self.instance else None,
-            'proposal': proposal,
-            'initial': {
-                # Default campus selected 'Louvain-la-Neuve' if exist
-                'campus': Campus.objects.filter(name='Louvain-la-Neuve').first(),
-                # Default language French
-                'language': language.find_by_code('FR'),
-                'container_type': container_type
-            } ,
-            'person': self.person
-        }
-
-    def get_context(self):
-        return {
-            'subtype': self.subtype,
-            'learning_unit_form': self.learning_unit_form,
-            'learning_unit_year_form': self.learning_unit_year_form,
-            'learning_container_year_form': self.learning_container_year_external_form,
-            'learning_unit_external_form': self.learning_unit_external_form
-        }
-
-    def save(self, commit=True):
-        academic_year = self.academic_year
-        learning_unit_yr = save_learning_unit_year(self.forms, academic_year, commit, None )
-        external = forms[LearningUnitExternalModelForm].save(
-            learning_unit_year=learning_unit_yr,
-            acronym=forms[LearningUnitExternalModelForm].instance.acronym,
-            credits=forms[LearningUnitExternalModelForm].instance.credits,
-            url=forms[LearningUnitExternalModelForm].instance.url,
-            entity=forms[LearningUnitExternalModelForm].instance.entity,
-            commit=commit
-        )
-        return external
-
-    def is_valid(self):
-        res = super().is_valid()
-        self.learning_unit_external_form.post_clean(self.academic_year.start_date)
-        return res
