+++ base/models/external_learning_unit_year.py
@@ -24,11 +24,10 @@
 #
 ##############################################################################
 
-from django.core.validators import MinValueValidator, MaxValueValidator, RegexValidator
+from django.core.validators import MinValueValidator, MaxValueValidator
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
-from base.models.learning_unit import LEARNING_UNIT_ACRONYM_REGEX_ALL
 from base.models.learning_unit_year import MINIMUM_CREDITS, MAXIMUM_CREDITS
 from base.models.osis_model_admin import OsisModelAdmin
 
@@ -44,12 +43,12 @@ class ExternalLearningUnitYear(models.Model):
     external_id = models.CharField(max_length=100, blank=True, null=True)
     changed = models.DateTimeField(null=True, auto_now=True)
     external_acronym = models.CharField(max_length=15, db_index=True, verbose_name=_('external_code'))
-    external_credits = models.DecimalField(max_digits=5, decimal_places=2,
+    external_credits = models.DecimalField(max_digits=5, decimal_places=2, verbose_name=_('local_credits'),
                                            validators=[MinValueValidator(MINIMUM_CREDITS),
                                                        MaxValueValidator(MAXIMUM_CREDITS)])
     url = models.URLField(max_length=255, blank=True, null=True)
     learning_unit_year = models.OneToOneField('LearningUnitYear')
-    buyer = models.ForeignKey('Entity')
+    buyer = models.ForeignKey('Entity', verbose_name=_('requesting_entity'))
 
     class Meta:
         unique_together = ('learning_unit_year', 'external_acronym',)
