+++ base/models/organization_address.py
@@ -24,7 +24,7 @@
 #
 ##############################################################################
 from django.db import models
-from django.contrib import admin
+
 from base.models.osis_model_admin import OsisModelAdmin
