+++ base/templates/learning_unit/blocks/create_form_learning_unit_external.html
@@ -27,25 +27,20 @@
 <div class="panel panel-default">
     <div class="panel-heading">
         <label>{% trans 'basic_informations_title' %}</label>
-       a {% bootstrap_form_errors learning_unit_year_form %}
+        {% bootstrap_form_errors learning_unit_year_form %}
         {% bootstrap_form_errors learning_unit_form %}
         {% bootstrap_form_errors learning_container_year_form %}
-        {% bootstrap_form_errors learning_unit_external_form %}b
+        {% bootstrap_form_errors learning_unit_external_form %}
     </div>
     <div class="panel-body">
         <div class="row">
-            <div class="col-md-3 acronym-group">
-                {% bootstrap_label learning_unit_year_form.acronym.label %}
-                <abbr title="{% trans 'acronym_rules' %}">
-                <span class="glyphicon glyphicon-question-sign " title="{% trans 'acronym_rules' %}" style="font-size: 14pt;"></span>
-                </abbr>
-                {% bootstrap_field learning_unit_year_form.acronym show_label=False %}
-            </div>
-            {% bootstrap_field learning_container_year_form.container_type form_group_class="col-md-3" %}
+            {% bootstrap_field learning_unit_year_form.acronym form_group_class="col-md-3 acronym-group"%}
+            <div class="col-md-5"></div>
+            {% bootstrap_field learning_container_year_form.container_type form_group_class="col-md-4" %}
         </div>
+        <br>
         <div class="row">
-            {% bootstrap_field learning_unit_year_form.academic_year form_group_class="col-md-3" %}
-            {% bootstrap_field learning_container_year_form.language form_group_class="col-md-3" %}
+            {% bootstrap_field learning_container_year_form.language form_group_class="col-md-6" %}
             {% bootstrap_field learning_unit_year_form.credits form_group_class="col-md-2" %}
             {% bootstrap_field learning_unit_external_form.external_credits form_group_class="col-md-2" %}
             {% bootstrap_field learning_unit_year_form.status form_group_class="col-md-1" %}
