+++ base/templates/learning_unit/simple/creation_external.html
@@ -49,7 +49,10 @@
                                 <label>{% trans 'institution' %}</label>
                             </div>
                             <div class="panel-body">
+{#                                {% bootstrap_field learning_container_year_form.country %}#}
                                 {% bootstrap_field learning_container_year_form.campus %}
+                                {% bootstrap_field learning_unit_year_form.academic_year %}
+
                             </div>
                         </div>
                     </div>
